Aug 12, 2024: Verbose ok

+++ include/global.h
@@ -14,7 +14,7 @@
 extern std::chrono::high_resolution_clock::time_point start;
 
 // flags are global variables
-extern short int tabular_flag;
+extern int tabular_flag;
 extern int verbose_flag;
 extern int seqReport_flag;
 extern int outSequence_flag;
